Move getGpuTags() to processcheck, use tagger
