+++ comp/process/processcheck/processcheckimpl/check.go
@@ -11,6 +11,7 @@ import (
 
 	"github.com/DataDog/datadog-agent/comp/core/config"
 	"github.com/DataDog/datadog-agent/comp/core/sysprobeconfig"
+	tagger "github.com/DataDog/datadog-agent/comp/core/tagger/def"
 	workloadmeta "github.com/DataDog/datadog-agent/comp/core/workloadmeta/def"
 	"github.com/DataDog/datadog-agent/comp/process/processcheck"
 	"github.com/DataDog/datadog-agent/comp/process/types"
@@ -36,6 +37,7 @@ type dependencies struct {
 	Config    config.Component
 	Sysconfig sysprobeconfig.Component
 	WMmeta    workloadmeta.Component
+	Tagger    tagger.Component
 }
 
 type result struct {
@@ -47,7 +49,7 @@ type result struct {
 
 func newCheck(deps dependencies) result {
 	c := &check{
-		processCheck: checks.NewProcessCheck(deps.Config, deps.Sysconfig, deps.WMmeta),
+		processCheck: checks.NewProcessCheck(deps.Config, deps.Sysconfig, deps.WMmeta, deps.Tagger),
 	}
 	return result{
 		Check: types.ProvidesCheck{
