+++ pkg/process/checks/checks.go
@@ -10,6 +10,7 @@ import (
 	model "github.com/DataDog/agent-payload/v5/process"
 
 	sysconfigtypes "github.com/DataDog/datadog-agent/cmd/system-probe/config/types"
+	tagger "github.com/DataDog/datadog-agent/comp/core/tagger/def"
 	workloadmeta "github.com/DataDog/datadog-agent/comp/core/workloadmeta/def"
 	"github.com/DataDog/datadog-agent/comp/networkpath/npcollector"
 	"github.com/DataDog/datadog-agent/pkg/config/env"
@@ -107,9 +108,9 @@ func (p CombinedRunResult) RealtimePayloads() []model.MessageBody {
 // All is a list of all runnable checks. Putting a check in here does not guarantee it will be run,
 // it just guarantees that the collector will be able to find the check.
 // If you want to add a check you MUST register it here.
-func All(config, sysprobeYamlCfg pkgconfigmodel.ReaderWriter, syscfg *sysconfigtypes.Config, wmeta workloadmeta.Component, npCollector npcollector.Component) []Check {
+func All(config, sysprobeYamlCfg pkgconfigmodel.ReaderWriter, syscfg *sysconfigtypes.Config, wmeta workloadmeta.Component, tagger tagger.Component, npCollector npcollector.Component) []Check {
 	return []Check{
-		NewProcessCheck(config, sysprobeYamlCfg, wmeta),
+		NewProcessCheck(config, sysprobeYamlCfg, wmeta, tagger),
 		NewContainerCheck(config, wmeta),
 		NewRTContainerCheck(config, wmeta),
 		NewConnectionsCheck(config, sysprobeYamlCfg, syscfg, wmeta, npCollector),
